use as few registers as possible in cuda kernels
Register usage was optimized to use 32 bit integers whenever possible.
However, it relied on the assumption that the maximum size was 32
relations.
This patch replaces all these 32 bit integers with a type that is inferred
from the maximum size of the bitmapset.

+++ src/backend/optimizer/gpuqo/gpuqo_dpsub_tree.cuh
@@ -63,7 +63,7 @@ static JoinRelation dpsubEnumerateTreeSimple(RelationID relid,
 
 __device__
 static JoinRelation dpsubEnumerateTreeWithSubtrees(RelationID relid, 
-                        uint64_t cid, int n_splits,
+                        uint32_t cid, int n_splits,
                         HashTableType &memo, GpuqoPlannerInfo* info)
 { 
     JoinRelation jr_out;
